#215 code cleanup
